rename organization and changes in pom.xml with artifact name and changed entity Company with add Lombok annotation
